// src/dashboard-next/src/lib/api.generated.ts
export type GeneratedDashboardOperationId =
  | "actOnCampaign"
  | "answerAgentQueriesBatch"
  | "answerAgentQuery"
  | "approveAction"
  | "approveActionsBatch"
  | "bundleEngagement"
  | "cancelAgent"
  | "cloneCampaign"
  | "closeSession"
  | "correctGraph"
  | "createCampaign"
  | "createEngagement"
  | "createEngagementFromTemplate"
  | "createObjective"
  | "deleteCampaign"
  | "deleteObjective"
  | "deleteReport"
  | "denyAction"
  | "denyActionsBatch"
  | "dismissAgent"
  | "dismissFleetAgents"
  | "dispatchAgent"
  | "dispatchAgentBatch"
  | "dispatchCampaign"
  | "downloadReport"
  | "explainAction"
  | "exportGraph"
  | "findPaths"
  | "getActionOutput"
  | "getActiveApplicationCommands"
  | "getAgentArchetypes"
  | "getAgentConsole"
  | "getAgentContext"
  | "getAgentDuplicates"
  | "getAgentHistory"
  | "getAgentQueries"
  | "getAgents"
  | "getApplicationCommand"
  | "getCampaign"
  | "getCampaignChildren"
  | "getConfig"
  | "getDecisionLog"
  | "getEngagement"
  | "getEvidenceChains"
  | "getEvidenceImage"
  | "getEvidenceRaw"
  | "getFindingContext"
  | "getFindings"
  | "getFrontierWeights"
  | "getGraph"
  | "getHealth"
  | "getHistory"
  | "getInferenceRules"
  | "getMcpTools"
  | "getObjectivePaths"
  | "getOperatorConsole"
  | "getOpsecBudget"
  | "getParsers"
  | "getPendingActions"
  | "getPhases"
  | "getPlaybookRun"
  | "getProposedPlans"
  | "getReadiness"
  | "getRecovery"
  | "getRuntime"
  | "getSessionBuffer"
  | "getSessions"
  | "getSettings"
  | "getState"
  | "getTapeStatus"
  | "getTelemetry"
  | "getTemplates"
  | "getTimeline"
  | "getTools"
  | "getTrustSignals"
  | "handoffAgent"
  | "interpretCommand"
  | "interruptPlaybookAttempt"
  | "issueAgentDirective"
  | "issueFleetDirective"
  | "listCampaigns"
  | "listEngagements"
  | "listPlaybookRuns"
  | "listReports"
  | "mergeAgent"
  | "previewScope"
  | "quickDeployAgent"
  | "renderReport"
  | "reparseAction"
  | "resetFrontierWeights"
  | "resolveConfigDivergence"
  | "resumePlaybookRun"
  | "resumeSession"
  | "retryPlaybookStep"
  | "skipPlaybookStep"
  | "splitAgent"
  | "splitCampaign"
  | "startPlaybookStep"
  | "toggleTape"
  | "updateCampaign"
  | "updateConfig"
  | "updateEngagement"
  | "updateFrontierWeights"
  | "updateObjective"
  | "updateScope"
  | "updateSession"
  | "updateSettings";

export interface GeneratedDashboardRequest {
  path?: Record<string, unknown>;
  query?: Record<string, unknown>;
  body?: unknown;
  headers?: HeadersInit;
  signal?: AbortSignal;
  cache?: RequestCache;
}
export type GeneratedDashboardRequestFor<T extends GeneratedDashboardOperationId> =
  Omit<GeneratedDashboardRequest, 'path' | 'query' | 'body'> & {
    path?: DashboardPathInput<T>;
    query?: DashboardQueryInput<T>;
    body?: DashboardBodyInput<T>;
  };
export type GeneratedDashboardOutput<T extends GeneratedDashboardOperationId> =
  DashboardEndpoint<T>['response_kind'] extends 'binary'
    ? Response
    : DashboardSuccessOutput<T>;

export class DashboardApiError extends Error {
  constructor(
    message: string,
    readonly status: number,
    readonly code?: string,
    readonly body?: unknown,
  ) {
    super(message);
    this.name = 'DashboardApiError';
  }
}